feat: Add CreateCardModal component
The CreateCardModal component is added to the codebase. This component allows for creating a new card within a deck. It provides a modal interface with input fields for the question and answer, as well as a submit button. This component enhances the user experience by providing a convenient way to add new cards to a deck.